Move decision making into a separate module
